+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Release 5.1.0:
    - `WalletService` supports building multiple authorization details to request a token for more than one credential
    - Remove `buildAuthorizationDetails(RequestOptions)` for `WalletService`, please migrate to `buildScope(RequestOptions)`
    - Note that multiple `scope` values may be joined with a whitespace ` `
+ - SD-JWT:
+   - Pass around decoded data with `SdJwtSigned` in several result classes like `VerifyPresentationResult.SuccessSdJwt`
+   - Rename `disclosures` to `validatedItems` in several result classes like `AuthnResponseResult.SuccessSdJwt`
 
 Release 5.0.1:
  - Update JsonPath4K to 2.4.0